Define solver template instantiations in their own files
