+++ src/NBody/Simulation/CMakeLists.txt
@@ -40,7 +40,7 @@ add_library(Simulation
         Runners/ContinuousRunner.cpp Runners/ContinuousRunner.h
         Solvers/LinearBVHSolver.h
         Solvers/DualTreeSolver.h
-        Solvers/DualTraversalSolver.h Solvers/MVDRSolver.h Solvers/Trees/Summaries/BoundingBoxSummary.h Solvers/DualTraversalSolverBase.h)
+        Solvers/DualTraversalSolver.h Solvers/MVDRSolver.h Solvers/Trees/Summaries/BoundingBoxSummary.h Solvers/DualTraversalSolverBase.h Solvers/ActiveTreeSolver.h Solvers/OctreeDualTraversalSolver.h)
 
 target_include_directories(Simulation PUBLIC
         ${CMAKE_SOURCE_DIR}/src)
